requirements.txt
# local package
-e .

# external requirements
click~=8.1.3
python-dotenv>=0.5.1

setuptools~=60.2.0
pandas~=1.4.4
src~=0.1.0
numpy~=1.23.3
scikit-learn~=1.0.2
joblib~=1.1.0
plotly~=5.10.0
requests~=2.28.1